rename: move RenteeInfo from info-list-layout to components
